fix: do some tidy up
