+++ apps/backend/package.json
@@ -102,11 +102,11 @@
         "pino-http": "^11.0.0",
         "pino-opentelemetry-transport": "^3.0.0",
         "pino-pretty": "^13.0.0",
+        "pkcs11js": "^2.1.6",
         "reflect-metadata": "^0.2.2",
         "rxjs": "^7.8.2",
         "sqlite3": "^5.1.7",
         "typeorm": "^0.3.28",
-        "pkcs11js": "^2.1.6",
         "uuid": "^14.0.0"
     },
     "devDependencies": {
@@ -123,6 +123,7 @@
         "@types/express": "^5.0.6",
         "@types/multer": "^2.0.0",
         "@types/node": "^25.3.3",
+        "@types/passport-jwt": "^4.0.1",
         "@types/supertest": "^7.2.0",
         "@vitest/coverage-v8": "^4.0.18",
         "dotenv": "^17.3.1",
