+++ apps/backend/tsconfig.json
@@ -14,7 +14,7 @@
         "skipLibCheck": true,
         "strictNullChecks": true,
         "forceConsistentCasingInFileNames": true,
-        "noImplicitAny": false,
+        "noImplicitAny": true,
         "strictBindCallApply": false,
         "noFallthroughCasesInSwitch": false
     }
